git-svn-id: https://svn.r-project.org/R/branches/ALTREP@73411 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -3706,7 +3706,7 @@ attribute_hidden
 size_t Rconn_getline(Rconnection con, char *buf, size_t bufsize)
 {
     int c;
-    size_t nbuf = -1;
+    ssize_t nbuf = -1;
 
     while((c = Rconn_fgetc(con)) != R_EOF) {
 	if(nbuf+1 >= bufsize)
@@ -3726,7 +3726,7 @@ size_t Rconn_getline(Rconnection con, char *buf, size_t bufsize)
 	    error(_("line longer than buffer size %lu"), (unsigned long) bufsize);
 	buf[++nbuf] = '\0';
     }
-    return(nbuf);
+    return (size_t) nbuf;
 }
 
 int Rconn_printf(Rconnection con, const char *format, ...)
